\name{NAEP}

\docType{data}

\alias{NAEP}

\title{National Assessment of Educational Progress (NAEP) p-values}

\description{P-values for the null hypothesis of no change in the average eighth-grade mathematics achievement scores for the 34 states that participated in both the 1990 and the 1992 National Assessment of Educational Progress (NAEP) Trial State Assessment.}

\usage{data(NAEP)}

\format{A named numeric vector of p-values, uncorrected for multiple testing.}

\references{Benjamini Y. and Hochberg, Y. 2000. On the adaptive control of the false discovery rate in multiple testing with independent statistics. Journal of Educational and Behavioral Statistics 25 (1) 60-83.

Williams, V.S.L., Jones, L.V. and Tukey, J.W. 1999. Controlling error in multiple
comparisons, with examples from state-to-state differences in educational achievement. Journal of Educational and Behavioral Statistics 24 (1) 42-69.}

\keyword{datasets}
